Minor updates
Updated comments to be more readable, corrected some typos.
Minor refactoring with regard to variable naming in the auto-refresh
class.  Changed signatures for auto-updating to be more readable.
Changed one .format to an f-string. Switched all classes to use slots.
Also decided to use the := operator to avoid multiple assignment
statements in a while loop